FIX: move gameboard to reinforcement folder and fix condition in MCTS
